Create ClosedOpportunityTrigger
Created a trigger for the Opportunity object. This trigger adds a task to any opportunity inserted or updated, with the stage of 'Closed Won'. The task's subject is 'Follow-Up Test Task'. To associate the task with the opportunity, fill the 'WhatId' field with the opportunity ID. This code tests 200 records in one operation.