Create CreateContact
created a VisualForce page in Apex that inserts a Contact record based on First Name, Last Name, and Email. After submitting the form, the user will be redirected to the detail page of the new Contact record. Page references the Contact standard controller, and uses a VisualForce apex:form component.