Create TestVerifyDate
Apex class to test is called "VerifyDate", which tests if a date is in a proper range, and if not, won't return a date that occurs at the end of the month within the range. Test unit covers scenarios for all lines of code in the Apex class, resulting in 100% code coverage.